Restructure project: Move server.py to wireshark_mcp package
