+++ tests/test_capture_flow.py
@@ -9,7 +9,7 @@ import sys
 ROOT = Path(__file__).resolve().parent.parent
 if str(ROOT) not in sys.path:
     sys.path.insert(0, str(ROOT))
-from server import perform_live_capture_enhanced
+from wireshark_mcp.server import perform_live_capture_enhanced
 
 
 class FakeProcess:
