+++ tests/test_filters.py
@@ -8,7 +8,7 @@ ROOT = Path(__file__).resolve().parent.parent
 if str(ROOT) not in sys.path:
     sys.path.insert(0, str(ROOT))
 
-from server import advanced_filter_generation, combine_filters_intelligently
+from wireshark_mcp.server import advanced_filter_generation, combine_filters_intelligently
 
 
 def test_combine_filters_and_logic():
