+++ tests/test_parsers.py
@@ -5,7 +5,7 @@ ROOT = Path(__file__).resolve().parent.parent
 if str(ROOT) not in sys.path:
     sys.path.insert(0, str(ROOT))
 
-from server import (
+from wireshark_mcp.server import (
     parse_protocol_hierarchy,
     parse_conversations,
     parse_endpoints,
