+++ tests/test_registry.py
@@ -5,7 +5,7 @@ ROOT = Path(__file__).resolve().parent.parent
 if str(ROOT) not in sys.path:
     sys.path.insert(0, str(ROOT))
 
-from server import list_tools, server
+from wireshark_mcp.server import list_tools, server
 
 
 def test_tools_registry_contains_expected():
