+++ wireshark_mcp/__init__.py
@@ -1,5 +1,5 @@
 # Convenience re-exports from top-level server module
-from server import (
+from wireshark_mcp.server import (
     call_tool,
     handle_realtime_json_capture,
     handle_protocol_statistics,
